requirements.txt
google-generativeai==0.3.2
streamlit==1.29.0
Pillow==10.1.0
python-dotenv==1.0.0
aiosignal>=1.1.2
frozenlist>=1.1.1
multidict>=4.5,<7.0
yarl>=1.0,<2.0
async-timeout>=4.0.0
cryptography>=2.0.0,<45.0.0
speech_recognition==3.10.0
gTTS==2.3.2
opencv-python-headless==4.8.1.78
streamlit-webrtc==0.47.1
playsound==1.3.0
